base32: move from src/ to external/
Similar to bignum.
