+++ .gitignore
@@ -23,6 +23,7 @@ external/*
 !external/ctaes
 !external/sha3
 !external/bignum
+!external/base32
 !external/CMakeLists.txt
 src/generated
 py/bitbox02/generated
