+++ external/base32/base32.h
